refactor: Extract FormRepository.
